requirements.txt
pyrogram[fast]
requests
TgCrypto
bs4
clint
python-magic
timeago
python-telegram-bot==12.0.0b1
